+++ src/main/java/com/github/openplay/model/CampaignStatesInterface.java
@@ -2,9 +2,9 @@ package com.github.openplay.model;
 
 public interface CampaignStatesInterface {
 	
-	public Long getId();
+	public int getId();
 	
-	public void setId(Long campaign_StateId);
+	public void setId(int campaign_StateId);
 	
 	public String getDescription();
