+++ src/main/java/com/github/openplay/model/CampaignTypeInterface.java
@@ -2,9 +2,9 @@ package com.github.openplay.model;
 
 public interface CampaignTypeInterface {
 
-		public Long getId();
+		public int getId();
 
-		public void setId(Long id);
+		public void setId(int id);
 
 		public String getDescription();
